refactored flask rest api structure
