+++ backend/database/models.py
@@ -1,0 +1,7 @@
+from .db import db
+
+# Example DB Object Schema
+# class Movie(db.Document):
+#     name = db.StringField(required=True, unique=True)
+#     casts = db.ListField(db.StringField(), required=True)
+#     genres = db.ListField(db.StringField(), required=True)
