+++ backend/requirements.txt
@@ -8,9 +8,12 @@ click==7.1.2
 colorama==0.4.1
 cycler==0.10.0
 distlib==0.3.0
+dnspython==2.0.0
 filelock==3.0.12
 Flask==1.1.2
+flask-mongoengine==0.9.5
 Flask-RESTful==0.3.8
+Flask-WTF==0.14.3
 idna==2.10
 importlib-metadata==1.6.0
 isort==4.3.21
@@ -21,12 +24,15 @@ lazy-object-proxy==1.4.1
 MarkupSafe==1.1.1
 matplotlib==3.2.2
 mccabe==0.6.1
+mongoengine==0.20.0
 mysql-connector-python==8.0.21
 numpy==1.19.0
 pandas==1.0.5
 pipenv==2018.11.26
+protobuf==3.13.0
 pycodestyle==2.6.0
 pylint==2.3.1
+pymongo==3.11.0
 pyparsing==2.4.7
 python-dateutil==2.8.1
 pytz==2020.1
@@ -39,4 +45,5 @@ virtualenv==20.0.20
 virtualenv-clone==0.5.4
 Werkzeug==1.0.1
 wrapt==1.11.2
+WTForms==2.3.3
 zipp==3.1.0
